requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.0
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.65.0
urllib3==2.0.4
watchdog==3.0.0
Werkzeug==2.3.6
yarl==1.9.2
